Make InvalidParameterValue exceptions more informative.
Add an additional text parameter to InvalidParameterValue to express
ExceptionText.  This is helpful for indicating the value of the invalid
parameter (see OGC 06-121r9, Table 27, Footnote a).  Utilize this parameter
throughout the code base where it makes sense.

Also fix Execute template so that ExceptionReport tag is correctly in the
ows namespace with a version attribute scpecified.  Without this, schema
validation fails.