Replace the current management of ExecuteResponse.status
The current status management is a custom list of status that also
define the behaviour of the storage of the status file. This newer
version of status management drop the old scheme and reuse the status
defined by WPS standard, i.e. accepted, started, succeeded, and failed.
It also separate the management of behaviour of the status file.