Fix deserialization (#456)
* when deserializing the WPSRequest, the inputs should be from ...

pywps.inout.inputs instead of pywps.inout.basic

* fixes for deserialization of a complex input of type file or url

* tests: remove __init__.py in requests folder containng static xml files

This folder was getting confused with the `requests` module

* fix error message missing description and uniformize messages

* move json serialization methods from WPSRequest.json to inout.inputs...

I put them in inout.inputs instead of inout.basic because
ComplexInput.json was already there and json methods for outputs are in
inout.outputs, not in inout.basic

* fix a circular dependency between inout.formats and ...

validator.complexvalidator. In all cases where Format is instantiated,
the `validate` property is set. So this shouldn't break existing code.

I prefer this solution to having a delayed import...

* add __eq__ mathods for testing values comparison

* json method has moved from inout.basic to inout.inputs

* fix handling and validation of allowed_values property

* add workdir to docstrings

* the `max_size` property of the ComplexInput is always queried from...

the global settings. It's not managed at a per-instance level.
So I removed the property and the unused method that was fetching
it from the configuration.

* add missing metadata to ComplexInput serialization

* handle serialization of 'file', 'data', 'stream' and 'url' ...

ComplexInput types.

* unused imports

* add serialization tests for ComplexInputs, LiteralInput and ...

BoundingBoxInput

* add title property. The serialization tests should really test the ...

classes in inout.inputs, as they are the classes used by the api.

* `uom` doesn't need to be serialized, because `uoms` is

* the AnyValue property should be kept when creating an input from json.

This makes the serialization code simpler, as we don't have to check
for the any_value flag, like it does currently...

* typo

* serialize both 'uoms' and 'uom' in LiteralInput because there can ...

be more than one

* add missing metadata in LiteralInput serialization

* add expected CDATA tag

* add missing 'workdir' parameter passed to DataHandler.__init__

* remove redundant comment

* BoundingBoxOutput should inherit from basic.BBoxOutput, not BBoxInput

* add json serialization to BoundingBoxOutput

* add simple docstrings

* add Metadata serialization

* fix redefinition of data variable