Implement and use dblog.cleanup_crashed_process
The dblog.cleanup_crashed_process implement an heuristic to detect
crashed sub-process, i.e. process that did not terminated in the
expected way. This may happen if user use kill with SIGKILL or the
process get killed by OOM killer or the process exit by himself with
exit(0) or some case that I do not have in mind. In this case the
process entry in the database is not updated properly to reflect the
fact that the process has finished. This process entry stack up and
block new request to be processed. This function prevent this by
checking if a process with the same pid still alive, if not process with
the same pid is found we are sure that the process is not running
anymore, thus we update is status as failed.

Currently this function is called when the limit of parralle request is
reach to check if some process are crashed.