Implement DetachProcessing
DetachProcesing intended to replace the broken MultiProcessing. Detach
process will create a sub-process that will be detached from it's parent
process, i.e. the parent pid will be 1. That way prevent sub-process to
stay in zombie mode, as the default MultiProcessing do.

This mode is crafted for linux platform. To use it add the following
configuration:

[processing]
mode=detachprocessing